added my topics page

+++ apps/api/src/controllers/topicController.ts
@@ -13,14 +13,16 @@ import {
   StatusCodes,
   UpdateTopicResponse,
 } from "@mcqapp/types";
+import { CustomRequest } from "../middlewares";
 
-const createTopicQuery = `INSERT INTO topics (title) VALUES ($1) RETURNING *;`;
+const createTopicQuery = `INSERT INTO topics (title, created_by) VALUES ($1, $2) RETURNING *;`;
 export const CreateTopic = asyncErrorHandler(async (req, res) => {
   const { success, data } = CreateTopicInput.safeParse(req.body);
+  const { user_id } = req as CustomRequest;
   if (!success)
     throw new CustomError("Invalid Input", StatusCodes.InvalidInput);
 
-  const result = await client.query(createTopicQuery, [data.title]);
+  const result = await client.query(createTopicQuery, [data.title, user_id]);
 
   const response: CreateTopicResponse = {
     message: "Topic Created Successfully",
@@ -31,14 +33,18 @@ export const CreateTopic = asyncErrorHandler(async (req, res) => {
   res.json(response);
 });
 
-const updateTopicQuery = `UPDATE topics SET title=$1 WHERE topic_id=$2;`;
+const updateTopicQuery = `UPDATE topics SET title=$1 WHERE topic_id=$2 AND created_by=$3;`;
 export const UpdateTopic = asyncErrorHandler(async (req, res) => {
   const { success, data } = UpdateTopicInput.safeParse(req.body);
-
+  const { user_id } = req as CustomRequest;
   if (!success)
     throw new CustomError("Invalid Input", StatusCodes.InvalidInput);
 
-  await client.query(updateTopicQuery, [data.new_title, data.topic_id]);
+  await client.query(updateTopicQuery, [
+    data.new_title,
+    data.topic_id,
+    user_id,
+  ]);
 
   const response: UpdateTopicResponse = {
     message: "Topic Updated Successfully",
@@ -48,14 +54,14 @@ export const UpdateTopic = asyncErrorHandler(async (req, res) => {
   res.json(response);
 });
 
-const deleteTopicQuery = `DELETE FROM topics WHERE topic_id=$1;`;
+const deleteTopicQuery = `DELETE FROM topics WHERE topic_id=$1 AND created_by=$2;`;
 export const DeleteTopic = asyncErrorHandler(async (req, res) => {
   const { success, data } = DeleteTopicInput.safeParse(req.body);
-
+  const { user_id } = req as CustomRequest;
   if (!success)
     throw new CustomError("Invalid Input", StatusCodes.InvalidInput);
 
-  await client.query(deleteTopicQuery, [data.topic_id]);
+  await client.query(deleteTopicQuery, [data.topic_id, user_id]);
 
   const response: DeleteTopicResponse = {
     message: "Topic Deleted Successfully",
@@ -65,9 +71,10 @@ export const DeleteTopic = asyncErrorHandler(async (req, res) => {
   res.json(response);
 });
 
-const getAllTopics = `SELECT * FROM topics;`;
+const getAllTopics = `SELECT * FROM topics WHERE created_by = ANY($1);`;
 export const GetTopics = asyncErrorHandler(async (req, res) => {
-  const result = await client.query(getAllTopics);
+  const { user_id } = req as CustomRequest;
+  const result = await client.query(getAllTopics, [[1, user_id]]);
 
   const response: GetTopicsResponse = {
     message: "All Topics",
